+++ lib/pow/ecto/schema/changeset.ex
@@ -87,6 +87,7 @@ defmodule Pow.Ecto.Schema.Changeset do
     |> Changeset.validate_required([:password_hash])
   end
 
+  # TODO: Remove `confirm_password` support by 1.1.0
   @doc """
   Validates the confirm password field.
 
@@ -95,14 +96,46 @@ defmodule Pow.Ecto.Schema.Changeset do
   `nil`.
   """
   @spec confirm_password_changeset(Ecto.Schema.t() | Changeset.t(), map(), Config.t()) :: Changeset.t()
-  def confirm_password_changeset(user_or_changeset, params, _config) do
-    changeset = Changeset.cast(user_or_changeset, params, [:password, :confirm_password])
+  def confirm_password_changeset(user_or_changeset, %{confirm_password: password_confirmation} = params, _config) do
+    params =
+      params
+      |> Map.delete(:confirm_password)
+      |> Map.put(:password_confirmation, password_confirmation)
+
+    do_confirm_password_changeset(user_or_changeset, params)
+  end
+  def confirm_password_changeset(user_or_changeset, %{"confirm_password" => password_confirmation} = params, _config) do
+    params =
+      params
+      |> Map.delete("confirm_password")
+      |> Map.put("password_confirmation", password_confirmation)
+
+    convert_confirm_password_param(user_or_changeset, params)
+  end
+  def confirm_password_changeset(user_or_changeset, params, _config),
+    do: do_confirm_password_changeset(user_or_changeset, params)
+
+  # TODO: Remove by 1.1.0
+  defp convert_confirm_password_param(user_or_changeset, params) do
+    IO.warn("warning: passing `confirm_password` value to `#{inspect unquote(__MODULE__)}.confirm_password_changeset/3` has been deprecated, please use `password_confirmation` instead")
+
+    changeset = do_confirm_password_changeset(user_or_changeset, params)
+    errors    = Enum.map(changeset.errors, fn
+      {:password_confirmation, error} -> {:confirm_password, error}
+      error                           -> error
+    end)
+
+    %{changeset | errors: errors}
+  end
+
+  defp do_confirm_password_changeset(user_or_changeset, params) do
+    changeset = Changeset.cast(user_or_changeset, params, [:password])
 
     changeset
     |> Changeset.get_change(:password)
     |> case do
-      nil      -> changeset
-      password -> validate_confirm_password(changeset, password)
+      nil       -> changeset
+      _password -> Changeset.validate_confirmation(changeset, :password, required: true)
     end
   end
 
@@ -208,15 +241,6 @@ defmodule Pow.Ecto.Schema.Changeset do
     password_max_length = Config.get(config, :password_max_length, @password_max_length)
 
     Changeset.validate_length(changeset, :password, min: password_min_length, max: password_max_length)
-  end
-
-  defp validate_confirm_password(changeset, password) do
-    confirm_password = Changeset.get_change(changeset, :confirm_password)
-
-    case password do
-      ^confirm_password -> changeset
-      _                 -> Changeset.add_error(changeset, :confirm_password, "does not match confirmation", validation: :confirmation)
-    end
   end
 
   defp maybe_put_password_hash(%Changeset{valid?: true, changes: %{password: password}} = changeset, config) do